Encapsulated Storage service logic in Files module
