+++ CustomCADs.Production.sln
@@ -89,6 +89,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Files.Persistenc
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Files.Application", "src\Modules\Files\Application\CustomCADs.Files.Application.csproj", "{0A8BC426-18EE-40EC-B2C1-49B6D78B30DC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Files.Infrastructure", "src\Modules\Files\Infrastructure\CustomCADs.Files.Infrastructure.csproj", "{EC65ABD0-332D-435E-8F86-BC6F1702D175}"
+EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Idempotency", "Idempotency", "{F0B6DE86-B080-2A94-AC3D-E478B2AFB247}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Idempotency.Domain", "src\Modules\Idempotency\Domain\CustomCADs.Idempotency.Domain.csproj", "{8AFE5655-4D17-409D-AE62-243940D5D157}"
@@ -539,6 +541,18 @@ Global
 		{0A8BC426-18EE-40EC-B2C1-49B6D78B30DC}.Release|x64.Build.0 = Release|Any CPU
 		{0A8BC426-18EE-40EC-B2C1-49B6D78B30DC}.Release|x86.ActiveCfg = Release|Any CPU
 		{0A8BC426-18EE-40EC-B2C1-49B6D78B30DC}.Release|x86.Build.0 = Release|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Debug|x64.Build.0 = Debug|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Debug|x86.Build.0 = Debug|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Release|x64.ActiveCfg = Release|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Release|x64.Build.0 = Release|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Release|x86.ActiveCfg = Release|Any CPU
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175}.Release|x86.Build.0 = Release|Any CPU
 		{8AFE5655-4D17-409D-AE62-243940D5D157}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8AFE5655-4D17-409D-AE62-243940D5D157}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8AFE5655-4D17-409D-AE62-243940D5D157}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -802,6 +816,7 @@ Global
 		{EABD8F12-4FB1-4EAC-9B61-B8BD9E4483D5} = {C554CFC4-74AA-45C8-A62C-C714AE2E6A16}
 		{A04F3EE1-26DA-4877-A522-A74D5533B606} = {C554CFC4-74AA-45C8-A62C-C714AE2E6A16}
 		{0A8BC426-18EE-40EC-B2C1-49B6D78B30DC} = {C554CFC4-74AA-45C8-A62C-C714AE2E6A16}
+		{EC65ABD0-332D-435E-8F86-BC6F1702D175} = {C554CFC4-74AA-45C8-A62C-C714AE2E6A16}
 		{F0B6DE86-B080-2A94-AC3D-E478B2AFB247} = {8CF3001A-FDE9-4E6A-A642-83480DB4B117}
 		{8AFE5655-4D17-409D-AE62-243940D5D157} = {F0B6DE86-B080-2A94-AC3D-E478B2AFB247}
 		{DEB02B2C-8303-4C57-B86C-7456913B1A85} = {F0B6DE86-B080-2A94-AC3D-E478B2AFB247}
