+++ src/Shared/Infrastructure/CustomCADs.Shared.Infrastructure.csproj
@@ -8,11 +8,9 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="AWSSDK.S3" />
 		<PackageReference Include="FluentEmail.Smtp" />
 		<PackageReference Include="MailKit" />
 		<PackageReference Include="Microsoft.Extensions.Caching.Memory" />
-		<PackageReference Include="Microsoft.Extensions.Options" />
 		<PackageReference Include="Stripe.net" />
 		<PackageReference Include="System.IdentityModel.Tokens.Jwt" />
 		<PackageReference Include="WolverineFx" />
